moving to standard types and macros

+++ Toolkit/RSPiX/GREEN/BLiT/Rotate96.cpp
@@ -20,6 +20,7 @@
 #include <BLUE/System.h>
 
 #include <GREEN/BLiT/Blit.h>
+#include <GREEN/BLiT/BlitInternal.h>
 #include <GREEN/BLiT/RFont.h>
 #include <ORANGE/QuickMath/Fractions.h>
 
@@ -59,7 +60,7 @@ inline void _BlitRot(int16_t sDeg,int16_t sHeight, // = 2R + 1?
 	// all things start at this source location:
 	uint8_t* pP;
 
-	int32_t	lDen = long(sDstW) * sDstH; // compound fraction
+   int32_t	lDen = int32_t(sDstW) * sDstH; // compound fraction
 
 	// Calculate the sub pixel offset:
 	int32_t lLadNumX,lLadNumY,lRungNumX,lRungNumY;
@@ -89,8 +90,8 @@ inline void _BlitRot(int16_t sDeg,int16_t sHeight, // = 2R + 1?
 	// 32-bit version
 	int32_t lPX = lDen * sR,lPY = lDen * sR; // go into Denominator space
 	// Stay in denominator space!
-	lPX += long(COSQ[sDegP] * sR * lDen + 0.5 * lDen); // offset to pixel center
-	lPY += long(SINQ[sDegP] * sR * lDen  + 0.5 * lDen); // offset to pixel center
+   lPX += int32_t(COSQ[sDegP] * sR * lDen + 0.5 * lDen); // offset to pixel center
+   lPY += int32_t(SINQ[sDegP] * sR * lDen  + 0.5 * lDen); // offset to pixel center
 	// Now MUST convert to an asymettrical signed proper fraction:
 	rspDivModA64(lPX,lDen,lPX,lLadNumX);
 	rspDivModA64(lPY,lDen,lPY,lLadNumY);
@@ -109,10 +110,10 @@ inline void _BlitRot(int16_t sDeg,int16_t sHeight, // = 2R + 1?
 	// Find the signed vector length of the rungs, in compound
 	// fraction form...
 
-	int32_t lRungW = long(COSQ[sDegR] * sEdge * sDstH); // horizontal
-	int32_t lRungH = long(SINQ[sDegR] * sEdge * sDstH);
-	int32_t lLadW = long(COSQ[sDegL] * sEdge * sDstW);  // vertical
-	int32_t lLadH = long(SINQ[sDegL] * sEdge * sDstW);
+   int32_t lRungW = int32_t(COSQ[sDegR] * sEdge * sDstH); // horizontal
+   int32_t lRungH = int32_t(SINQ[sDegR] * sEdge * sDstH);
+   int32_t lLadW = int32_t(COSQ[sDegL] * sEdge * sDstW);  // vertical
+   int32_t lLadH = int32_t(SINQ[sDegL] * sEdge * sDstW);
 	// Convert from a point offset to true width and height:
 
 	lRungW += SGN3(lRungW); // 3 phase sign
@@ -469,8 +470,8 @@ inline int16_t rspClipMirrorDst(RImage* pimDst, // input:
 										int16_t &sClipR,
 										int16_t &sClipT,
 										int16_t &sClipB,
-										long	&lDstP,	// Including mirroring
-										long	&lDstPX	// Incl. Mirroring & pixDepth
+                              int32_t	&lDstP,	// Including mirroring
+                              int32_t	&lDstPX	// Incl. Mirroring & pixDepth
 										)
 	{
 	int16_t sMirrorX = 1,sMirrorY = 1; // direction flags...
@@ -582,7 +583,7 @@ int16_t rspAddRotationPadding(RImage* pimSrc,int16_t sHotX,int16_t sHotY)
 	lR = MAX(lR,SQR(sHotX - (pimSrc->m_sWidth - 1) ) + lHotYS ); // LL corner
 	lR = MAX(lR,lHotXS + lHotYS ); // UL corner
 
-	lR = long(0.999999 + sqrt(double(lR) * 2.0)); // round up
+   lR = int32_t(0.999999 + sqrt(double(lR) * 2.0)); // round up
 	// The sqrt2 factor is needed because the moving window must enclose the circle.
 
 	int16_t sSize = int16_t(1 + (lR << 1) ); // buffer = 2R + 1
